server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
PyJWT==2.8.0
Werkzeug==3.0.1
pycryptodome==3.19.0
requests==2.31.0
Python-dotenv==1.0.0